+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add support for PHP 8.2+ (#64)
 - Make sure that the tab title has the note title (#105)
+- Add tab key press support (insert 4 spaces) (#63)
 
 
 ## [1.0.0] - 2024-12-13